requirements.txt
flask
google-cloud-aiplatform
google-genai
pypdf
numpy
werkzeug
gunicorn